+++ lantai/api/__init__.py
@@ -27,6 +27,7 @@ from .routes_ui import router as routes_ui_router
 from .routes_work_items import router as routes_work_items_router
 from .routes_persona import router as routes_persona_router
 from .routes_scratchpad import router as routes_scratchpad_router
+from .routes_probing import router as routes_probing_router
 
 __all__ = [
     "routes_health_router",